backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "annotated>=0.0.2",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.128.8",
    "file>=0.3.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.34",
]
